requirements.txt
google-api-python-client==2.97.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.1.0
pandas==2.0.3
requests==2.31.0
huggingface-hub==0.19.4